+++ shared/src/main/scala/citeobj/package.scala
@@ -2,6 +2,7 @@ package edu.holycross.shot
 
 import scala.collection.mutable.ArrayBuffer
 import scala.collection.mutable.Map
+import edu.holycross.shot.cite._
 
 package object citeobj {
 
@@ -26,6 +27,42 @@ package object citeobj {
       rowMaps += headerToCell
     }
     rowMaps.toVector
+  }
+
+  /** True if propertyValue is valid for propertyType.
+  *
+  * @param propertyValue Value for a CITE Collection property.
+  * @param propertyDefe Definition for the property.
+  */
+  def validValue(propertyValue: Any, propertyDef: CitePropertyDef): Boolean = {
+    propertyDef.propertyType match {
+      case CtsUrnType =>  propertyValue match {
+        case u: CtsUrn => true
+        case _ => false
+      }
+      case Cite2UrnType =>    propertyValue match {
+        case u: Cite2Urn => true
+        case _ => false
+      }
+      case NumericType =>  propertyValue match {
+        case i: Int => true
+        case d: Double => true
+        case bg: BigDecimal => true
+        case _ => false
+      }
+      case BooleanType =>   propertyValue match {
+        case b: Boolean => true
+        case _ => false
+      }
+      case ControlledVocabType  =>   propertyValue match {
+        case s: String => if (propertyDef.vocabularyList.contains(s)) { true } else { false }
+        case _ => false
+      }
+      case StringType  =>   propertyValue match {
+        case s: String =>  true
+        case _ => false
+      }
+    }
   }
 
 }