+++ releases.md
@@ -1,11 +1,13 @@
 # `citeobj`: release notes
 
 
+**2.0.0**: Introduces the `CitePropertyImplementation`.
+
 **1.4.0**: Adds functionality to `ImageExtensions` to select extensions by protocol, or for individual collections.
 
 **1.3.1**:  Fixes bugs in parsing CEX of image extensions. and of catalogs of Cite Collections.
 
-**1.3.0**:  Add `ImageExtensions`.
+**1.3.0**:  Adds `ImageExtensions` class and object.
 
 **1.2.1**:  Better merge of working branches.
 
@@ -13,6 +15,6 @@
 
 **1.1.1**: Bug fixes in exposing shared code to ScalaJS environment.
 
-**1.1.0**: Support building a data collection from a string source in CEX format.
+**1.1.0**: Supports building a data collection from a string source in CEX format.
 
 **1.0.0**: initial release.  Supports building a CiteCollectionRepository in various ways, including from a String in CEX 1.1 format.  Alternatively, in the JVM subproject, XML catalogs can be used with delimited text data to create a repository.